requirements.txt
asgiref==3.2.7
boto3==1.13.15
botocore==1.16.15
certifi==2020.4.5.1
Django==3.0.6
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.10.0
mysqlclient==1.4.6
pkg-resources==0.0.0
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
s3transfer==0.3.3
schedule==0.6.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
